NCI-Agency/anet#3621: Move header and footer banners
